% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{getQError}
\alias{getQError}
\title{Get error for given Q-matrix}
\usage{
getQError(qmat, unique.scores, freq)
}
\arguments{
\item{qmat}{q-matrix}

\item{unique.scores}{unique scores}

\item{freq}{frequency of unique scores.}
}
\description{
Get error for given Q-matrix
}
\examples{

library(ltm)
data("WIRS") 
scores <- getUniqueAnswersFreqs(WIRS)
qmat <- fitQmat(WIRS, concepts.no = 2, qmat = matrix(1, ncol(WIRS), nrow = 2))$final.mat
getQError(qmat, scores$unique.scores, scores$freq)

}
