% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{getUniqueAnswersFreqs}
\alias{getUniqueAnswersFreqs}
\title{Create frequency of the answers patterns for given dataset.}
\usage{
getUniqueAnswersFreqs(scores)
}
\arguments{
\item{scores}{data.frame with scores.}
}
\value{
List with three elements:
\itemize{
 \item{"count.scores"}{data frame with all answers patterns with a frequency column.}
 \item{"unique.scores"}{data frame with patterns without frequencies.}
 \item{"freq"}{a vector with frequencies of given pattern.}
}
}
\description{
Create frequency of the answers patterns for given dataset.
}
\details{
This function mith be usefull for calculating q-error using \code{\link{getQError}} function.
}
\examples{
library(ltm)
data("WIRS") 
scores <- getUniqueAnswersFreqs(WIRS)

}
