% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitQmat}
\alias{fitQmat}
\title{Fit Q Matrix}
\usage{
fitQmat(concepts.no, scores, scores.test, verbose = FALSE)
}
\arguments{
\item{concepts.no}{number of concepts to fit}

\item{scores}{table containing the students scores}

\item{scores.test}{validation set.}

\item{verbose}{if true, some additional information will be printed.}
}
\description{
Fit Q Matrix
}
\examples{

}
